fairlens-output/requirements.txt
# FairLens AI v2.0 — Requirements
# Install: pip install -r requirements.txt

# Web framework
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-multipart==0.0.9

# Data
pandas==2.2.2
numpy>=1.22.4,<2.3
scipy==1.13.1

# ML (for real accuracy delta in debiaser)
scikit-learn>=1.4.0

# PDF generation
reportlab==4.2.2

# HTTP client (for LLM API calls)
httpx==0.27.2

# Config
python-dotenv==1.0.1

# Session store (Redis backend — optional)
redis==5.0.8

# Typing
pydantic==2.9.2

# Testing
pytest==8.3.3
pytest-asyncio==0.24.0
pytest-cov==5.0.0
